+++ src/main/java/com/grookage/qtrouper/core/models/QueueContext.java
@@ -19,11 +19,15 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.google.common.base.Strings;
 import com.grookage.qtrouper.core.exceptions.TrouperExceptions;
 import com.grookage.qtrouper.utils.SerDe;
-import lombok.*;
-
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 
 /**
  * @author koushik
@@ -35,29 +39,45 @@ import java.util.Map;
 @ToString
 @Builder
 @SuppressWarnings("unused")
-public class QueueContext implements Serializable{
+public class QueueContext implements Serializable {
+
     private String serviceReference;
     @Builder.Default
     private Map<String, Object> data = new HashMap<>();
+    @Builder.Default
+    private int messagePriority = 0;
 
-    public <T> void addContext(Class<T> klass, T value) {
-        addContext(klass.getSimpleName().toUpperCase(), value);
+    public <T> void addContext(Class<T> klass,
+                               T value) {
+        addContext(klass.getSimpleName()
+                .toUpperCase(), value);
     }
 
     @JsonIgnore
     public <T> T getContext(Class<T> tClass) {
-        return getContext(tClass.getSimpleName().toUpperCase(), tClass);
+        return getContext(tClass.getSimpleName()
+                .toUpperCase(), tClass);
     }
 
-    public <T> void addContext(String key, T value) {
-        if (Strings.isNullOrEmpty(key.toUpperCase()))
+    public <T> void addContext(String key,
+                               T value) {
+        if (Strings.isNullOrEmpty(key.toUpperCase())) {
             TrouperExceptions.illegalArgument("Invalid key for context data. Key cannot be null/empty");
+        }
         this.data.put(key.toUpperCase(), value);
     }
 
     @JsonIgnore
-    public <T> T getContext(String key, Class<T> tClass) {
+    public <T> T getContext(String key,
+                            Class<T> tClass) {
         final var value = this.data.get(key.toUpperCase());
-        return null == value ? null : SerDe.mapper().convertValue(value, tClass);
+        return null == value
+               ? null
+               : SerDe.mapper()
+                       .convertValue(value, tClass);
+    }
+
+    public void resetMessagePriority() {
+        this.setMessagePriority(0);
     }
 }