+++ src/main/java/com/grookage/qtrouper/core/rabbit/RabbitConfiguration.java
@@ -16,14 +16,13 @@
 package com.grookage.qtrouper.core.rabbit;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import java.util.List;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
-import java.util.List;
 
 /**
  * @author koushik
@@ -34,6 +33,7 @@ import java.util.List;
 @Builder
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class RabbitConfiguration {
+
     @NotNull
     @NotEmpty
     private List<RabbitBroker> brokers;