+++ CHANGELOG.md
@@ -1,8 +1,15 @@
+### Version 1.0.2
+
+- Added Priority Queue Support to Trouper
+- Ability to declare a queue with maxPriority
+- Retry Queues (Since they work with TTLs) don't operate with priority but by default cascade the priority to the MQ
+- Sideline Publishes can be with a Priority, since consumers can be configured on a sideline. 
+
 ### Version 1.0.0
 
 - Moved to com.grookage from io.grookage.
 - Moved to java11 and dw 2.
-- Added sonar analysis 
+- Added sonar analysis
 
 ### Version 0.0.2
 
@@ -11,9 +18,13 @@
 
 ### Version 0.0.1-1
 
-- Added default handling to QTrouper.handle in case the properties object goes missing or when the headers are not present.
-- Doing a minor, for this is a bug fix. 
+- Added default handling to QTrouper.handle in case the properties object goes missing or when the
+  headers are not present.
+- Doing a minor, for this is a bug fix.
 
 ## Impact
 
-If you are using trouper to publish messages and read off it, this won't impact you. But when you are publishing messages using another RMQ client or an adhoc script that pushes messages into the queue without the headers, required (that trouper would've organically added), you'll see this issue. 
+If you are using trouper to publish messages and read off it, this won't impact you. But when you
+are publishing messages using another RMQ client or an adhoc script that pushes messages into the
+queue without the headers, required (that trouper would've organically added), you'll see this
+issue.